+++ packages/snap-account-service/src/index.ts
@@ -3,6 +3,9 @@ export type {
   SnapAccountServiceActions,
   SnapAccountServiceConfig,
   SnapAccountServiceEvents,
+  SnapAccountServiceAccountBalancesUpdatedEvent,
+  SnapAccountServiceAccountTransactionsUpdatedEvent,
+  SnapAccountServiceAccountAssetListUpdatedEvent,
   SnapAccountServiceMessenger,
   SnapAccountServiceOptions,
 } from './SnapAccountService';